+++ src/KsWare.Presentation.Lite/(Common)/MultiBinding.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace KsWare.Presentation.Lite {
+	// TODO MultiBinding
+	// public class MultiBinding : {
+	// }
+}