+++ src/KsWare.Presentation.Lite/(Common)/PriorityBinding.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace KsWare.Presentation.Lite {
+	// TODO PriorityBinding
+	// public class PriorityBinding :  {
+	//
+	// }
+
+}