+++ src/KsWare.Presentation.Lite/MarkupExtensions/xmlns.cs
@@ -1,0 +1,5 @@
+﻿using System.Windows.Markup;
+
+// add KsWare.Presentation.Lite.MarkupExtensions to default namespace, so no prefix is required
+[assembly: XmlnsDefinition("http://schemas.microsoft.com/winfx/2006/xaml/presentation", "KsWare.Presentation.Lite.MarkupExtensions")]
+