+++ ChangeLog.md
@@ -1,6 +1,12 @@
 ﻿# ChangeLog
 KsWare.Presentation.Lite
 
+#### [2020-01-30]
+- Binding/BindingBase markup extension
+- DebugBindung markup extension
+- DebugHint markup extension
+- ObjectDebugExtensions
+
 #### [2020-01-27]
 - ViewLocator, IViewLocatorStrategy, ...
 - ViewModelPresenter, ViewModelViewConverter